Data.AsciiTable

Description

Let's make a table!

> let Just (Object o1) = decode "{\"foo\": \"bar\"}"
> let Just (Object o2) = decode "{\"baz\": 5}"
> let Just (Object o3) = decode "{\"oink\": true}"

> let slice1 = [[Just o1, Just o3], [Just o2, Nothing]]
> let slice2 = [[Nothing, Just o1]]

> pretty (makeTable ["object 1", "object 2"] [slice1, slice2, slice1])
+-----------+------------+
| object 1  | object 2   |
|           |            |
| baz foo   | foo   oink |
+===========+============+
|     "bar" |       True |
| 5.0       |            |
+-----------+------------+
|           | "bar"      |
+-----------+------------+
|     "bar" |       True |
| 5.0       |            |
+-----------+------------+

Documentation

data Table

An opaque data type with a Pretty instance, for printing to a console. Build a table with makeTable, and show it with the pretty-printing functions re-exported from this module.

type TableRow a = [Maybe a]

A single horizontal row of a Table, containing a list of TableElems. Each element in the row is visually separated from the next by a vertical line. Each row in the table must contain the same number of elements (however, any number of them can be Nothing).

type TableSlice a = [TableRow a]

A single horizontal slice of a Table, containing one or more TableRows. Each slice is visually separated from the next by a horizontal line.

class TableElem a where

The class of types that correspond to a single element of a Table. An instance for an aeson Object is provided by this library.

Methods

tableElemCells :: a -> HashMap Text Text

makeTable :: forall a. TableElem a => [Text] -> [TableSlice a] -> Table

Make a Table from a list of headers and a list of TableSlices, each of which contains a list of TableRows, each of which contain a list of TableElems. It is assumed that all dimensions align properly (e.g. each row contains the same number of elements, which is equal to the length of the list of headers).

Each vertically aligned element need not contain the same set of keys; for example, the table corresponding to

[ {"foo": "bar"}, {"baz": "qux"} ]

will simply look like

+-------------+
| foo   baz   |
+=============+
| "bar"       |
|       "qux" |
+-------------+

That is, each missing value is simply not displayed.

Re-exports

data Doc e :: * -> *

The abstract data type Doc represents pretty documents.

Doc is an instance of the Show class. (show doc) pretty prints document doc with a page width of 100 characters and a ribbon width of 40 characters.

show (text "hello" `above` text "world")

Which would return the string "hello\nworld", i.e.

hello
world

putDoc :: Doc e -> IO ()

The action (putDoc doc) pretty prints document doc to the standard output, with a page width of 100 characters and a ribbon width of 40 characters.

main :: IO ()
main = do{ putDoc (text "hello" <+> text "world") }

Which would output

hello world

hPutDoc :: Handle -> Doc e -> IO ()

(hPutDoc handle doc) pretty prints document doc to the file handle handle with a page width of 100 characters and a ribbon width of 40 characters.

main = do{ handle <- openFile "MyFile" WriteMode
         ; hPutDoc handle (vcat (map text
                           ["vertical","text"]))
         ; hClose handle
         }

class Pretty a where

The member prettyList is only used to define the instance Pretty a => Pretty [a]. In normal circumstances only the pretty function is used.

Minimal complete definition

pretty

Methods

pretty :: a -> Doc e

prettyList :: [a] -> Doc e

data SimpleDoc e :: * -> *

The data type SimpleDoc represents rendered documents and is used by the display functions.

The Int in SText contains the length of the string. The Int in SLine contains the indentation for that line. The library provides two default display functions displayS and displayIO. You can provide your own display function by writing a function from a SimpleDoc to your own output format.

Constructors

SFail

SEmpty

SChar !Char (SimpleDoc e)

SText !Int String (SimpleDoc e)

SLine !Int (SimpleDoc e)

SEffect e (SimpleDoc e)

renderPretty :: Float -> Int -> Doc e -> SimpleDoc e

This is the default pretty printer which is used by show, putDoc and hPutDoc. (renderPretty ribbonfrac width x) renders document x with a page width of width and a ribbon width of (ribbonfrac * width) characters. The ribbon width is the maximal amount of non-indentation characters on a line. The parameter ribbonfrac should be between 0.0 and 1.0. If it is lower or higher, the ribbon width will be 0 or width respectively.

renderCompact :: Doc e -> SimpleDoc e

(renderCompact x) renders document x without adding any indentation. Since no 'pretty' printing is involved, this renderer is very fast. The resulting output contains fewer characters than a pretty printed version and can be used for output that is read by other programs.

renderSmart :: Int -> Doc e -> SimpleDoc e

A slightly smarter rendering algorithm with more lookahead. It provides provide earlier breaking on deeply nested structures. For example, consider this python-ish pseudocode: fun(fun(fun(fun(fun([abcdefg, abcdefg]))))) If we put a softbreak (+ nesting 2) after each open parenthesis, and align the elements of the list to match the opening brackets, this will render with renderPretty and a page width of 20c as: fun(fun(fun(fun(fun([ | abcdef, | abcdef, ] ))))) | Where the 20c. boundary has been marked with |. Because renderPretty only uses one-line lookahead, it sees that the first line fits, and is stuck putting the second and third lines after the 20c mark. In contrast, renderSmart will continue to check the potential document up to the end of the indentation level. Thus, it will format the document as:

fun(                |
  fun(              |
    fun(            |
      fun(          |
        fun([       |
              abcdef,
              abcdef,
            ]       |
  )))))             |

Which fits within the 20c. mark. In addition, renderSmart uses this lookahead to minimize the number of lines printed, leading to more compact and visually appealing output. Consider this example using the same syntax as above: aaaaaaaaaaa([abc, def, ghi]) When rendered with renderPretty and a page width of 20c, we get: aaaaaaaaaaa([ abc , def , ghi ]) Whereas when rendered with renderSmart and a page width of 20c, we get: aaaaaaaaaaa( [abc, def, ghi])

displayS :: SimpleDoc e -> ShowS

(displayS simpleDoc) takes the output simpleDoc from a rendering function and transforms it to a ShowS type (for use in the Show class).

showWidth :: Int -> Doc -> String
showWidth w x   = displayS (renderPretty 0.4 w x) ""

displayIO :: Handle -> SimpleDoc e -> IO ()

(displayIO handle simpleDoc) writes simpleDoc to the file handle handle. This function is used for example by hPutDoc:

hPutDoc handle doc  = displayIO handle (renderPretty 0.4 100 doc)